Improve integration with AVA
Have the watcher ignore changes to the original TypeScript source files.
This way, the watcher will only react to changes to the TypeScript build
output.

Allow the watcher to resolve the rewritten path of a test file, so it
can track it as a dependency of the original TypeScript source file.
This way, when the build output changes, the watcher knows to re-run the
original test file.

Update file patterns to ignore *.d.ts files, as well as any files in the
build output.

Updated the ignored-by-watcher patterns to ignore source map files from
the build output.

Fixes #10, #7, #9.